ImageManager.py: kernel file for backup Vu+

+++ src/ImageManager.py
@@ -1056,10 +1056,13 @@ class ImageBackup(Screen):
 				slot = getCurrentImage()
 				print("[ImageManager] Stage 1: Making kernel image.")
 			if "bin" or "uImage" in self.KERNELFILE:
-				if hasKexec:
-	#				boot = "boot" if slot > 0 and slot < 4 else "dev/%s/%s"  %(self.MTDROOTFS, self.ROOTFSSUBDIR)
-					boot = "boot"
-					self.command = "dd if=/%s/%s of=%s/vmlinux.bin" % (boot, canMultiBoot[slot]["kernel"].rsplit("/", 1)[1], self.WORKDIR) if slot != 0 else "dd if=/dev/%s of=%s/vmlinux.bin" % (self.MTDKERNEL, self.WORKDIR)
+				if brand == "vuplus":
+					if hasKexec:
+#						boot = "boot" if slot > 0 and slot < 4 else "dev/%s/%s"  %(self.MTDROOTFS, self.ROOTFSSUBDIR)
+						boot = "boot"
+						self.command = "dd if=/%s/%s of=%s/vmlinux.bin" % (boot, canMultiBoot[slot]["kernel"].rsplit("/", 1)[1], self.WORKDIR) if slot != 0 else "dd if=/dev/%s of=%s/vmlinux.bin" % (self.MTDKERNEL, self.WORKDIR)
+					else:
+						self.command = "dd if=/dev/%s of=%s/vmlinux.bin" % (self.MTDKERNEL, self.WORKDIR)
 				else:
 					self.command = "dd if=/dev/%s of=%s/kernel.bin" % (self.MTDKERNEL, self.WORKDIR)
 			else:
@@ -1357,6 +1360,8 @@ class ImageBackup(Screen):
 
 			if "bin" or "uImage" in self.KERNELFILE and exists("%s/kernel.bin" % self.WORKDIR):
 				move("%s/kernel.bin" % self.WORKDIR, "%s/%s" % (self.MAINDEST, self.KERNELFILE))
+			elif exists("%s/vmlinux.bin" % self.WORKDIR):
+				move("%s/vmlinux.bin" % self.WORKDIR, "%s/%s" % (self.MAINDEST, self.KERNELFILE))
 			else:
 				move("%s/vmlinux.gz" % self.WORKDIR, "%s/%s" % (self.MAINDEST, self.KERNELFILE))
 
@@ -1387,7 +1392,7 @@ class ImageBackup(Screen):
 				fileout.write(line)
 
 			if brand == "vuplus":
-				if model == "vuzero":
+				if model in ("vuzero4k, vuuno4k"):
 					with open(self.MAINDEST + "/force.update", "w") as fileout:
 						line = "This file forces the update."
 						fileout.write(line)
